+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add map, area and landmark remote state
 - Add ig.ENTITY.DynamicPlatform remote state
+- Add ig.ENTITY.NPC remote state
 
 ### Fixed